+++ src/main/java/ru/khasang/cachoeira/view/mainwindow/properties/panes/ResourcePropertiesPane.java
@@ -1,4 +1,4 @@
-package ru.khasang.cachoeira.view.mainwindow.properties;
+package ru.khasang.cachoeira.view.mainwindow.properties.panes;
 
 public class ResourcePropertiesPane extends AbstractPropertiesPane {
 }
